# appCatalogue/api-docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "The API of the NFV and SDN Applications Catalogue Service.",
		"version": "1.0",
		"title": "SELFNET Applications Catalogue",
		"contact": {
			"name": "Giacomo Bernini",
			"url": "https://selfnet-5g.eu/",
			"email": "g.bernini@nextworks.it"
		},
		"license": {
			"name": "Apache License Version 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8081",
	"basePath": "/",
...
	"paths": {
...
		"/app-catalogue/config/vims": {
			"get": {
				"tags": ["config-rest-controller"],
				"summary": "APP Package VIM Instances Query",
				"operationId": "getVimsUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/Collection«Vim»"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"post": {
				"tags": ["config-rest-controller"],
				"summary": "APP Package VIM Instance Configuration",
				"operationId": "addVimUsingPOST",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"in": "body",
					"name": "vim",
					"description": "vim",
					"required": true,
					"schema": {
						"$ref": "#/definitions/Vim"
					}
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created",
						"schema": {
							"type": "string"
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/app-catalogue/config/vims/{vimName}": {
			"get": {
				"tags": ["config-rest-controller"],
				"summary": "APP Package VIM Instance Query",
				"operationId": "getVimUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "vimName",
					"in": "path",
					"description": "vimName",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/Vim"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"delete": {
				"tags": ["config-rest-controller"],
				"summary": "APP Package VIM Instance Deletion",
				"operationId": "deleteVimUsingDELETE",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "vimName",
					"in": "path",
					"description": "vimName",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Succes",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"204": {
						"description": "No Content"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
...
		"Collection«Vim»": {
			"type": "object"
		},
...
		"Vim": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"tenant": {
					"type": "string"
				},
				"type": {
					"type": "string",
					"enum": ["OPENSTACK", "OTHER"]
				},
				"url": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			}
		},
...
	}
}